Code to apply Self-Healing to all vehicles, aircraft and infantry
Added an overloaded constructor to Aircraft, Infantry and Unit types and their super class TechnoTypeClass.
Added in the respective if statement to the AI function for the selfhealing.
Modified the self healing logic for vehicles as it was hard coded for mammoth tanks only. It is now able to be applied to any unit via a boolean parameter. Updated the mammoth tank to use this flag. Currently all units are stock.